+++ src/main/java/org/elasticsearch/index/query/HasParentQueryParser.java
@@ -39,7 +39,7 @@ import java.io.IOException;
 import java.util.HashSet;
 import java.util.Set;
 
-import static org.elasticsearch.index.query.HasChildFilterParser.ensureNotDeleteByQuery;
+import static org.elasticsearch.index.query.QueryParserUtils.ensureNotDeleteByQuery;
 
 public class HasParentQueryParser implements QueryParser {
